# week4-auth/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "W4 Auth API — Supabase Authentication",
    "description": "Week 4 Assignment — Secure API using Supabase Auth, JWTs, middleware protection, and Swagger UI.\n\n**How to test protected routes:**\n1. Use `POST /auth/login` to get an `access_token`\n2. Click the **Authorize 🔒** button above\n3. Paste your token (without 'Bearer ') and click **Authorize**\n4. All protected routes will now send your token automatically",
    "version": "1.0.0",
    "contact": {
      "name": "Flyrankai Intern Projects",
      "url": "https://github.com/kainatzahra98/FlyrankAI-intern-projects"
    }
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local development server" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Paste your Supabase access_token here (obtained from POST /auth/login)"
      }
    },
    "schemas": {
      "AuthCredentials": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email":    { "type": "string", "format": "email", "example": "user@example.com" },
          "password": { "type": "string", "minLength": 6,    "example": "password123" }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Authenticate and receive a JWT access token",
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AuthCredentials" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful — copy the access_token to authorize",
            "content": {
              "application/json": {
                "example": {
                  "message": "Login successful.",
                  "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                  "refresh_token": "some-refresh-token",
                  "expires_in": 3600,
                  "user": { "id": "uuid", "email": "user@example.com" }
                }
              }
            }
          },
          "400": { "description": "Missing fields" },
          "401": {
            "description": "Invalid credentials",
            "content": { "application/json": { "example": { "error": "Invalid login credentials" } } }
          }
        }
      }
    },
...
  }
}